refactor(schemas): migrate Phase 10 endpoint schemas — 6 files, 41 classes (#6042)
- heartbeat.py: migrate HeartbeatConfigRequest/Response, WakeupRequestCreate/Response,
  RunEventResponse, HeartbeatRunResponse → schemas_system.py
- advanced_control.py: migrate StreamingSessionRequest/Response, TakeoverRequest,
  TakeoverApprovalRequest, TakeoverActionRequest, SystemMonitoringResponse → schemas_system.py
- vision.py: migrate ScreenAnalysisRequest/Response, ElementDetectionRequest,
  OCRRequest, ElementInteractionRequest, UIElementResponse, VisionHealthResponse → schemas_system.py
- personality.py: migrate ProfileSummary/Detail/Create/Update, ToggleRequest,
  StatusResponse → schemas_agent.py with Personality* prefix
- chat.py: migrate ChatMessage, ChatResponse, MessageHistory, EnhancedChatMessage,
  ChatPreferences, TranslateRequest, DetectLanguageRequest → schemas_chat.py
- code_intelligence.py: migrate AnalysisRequest, QuickScanRequest, SuggestionsRequest
  → schemas_code.py with CodeIntel* prefix; migrate Redis*, Security*, Performance*
  AnalysisRequest/FileScanRequest as-is
