+++ autobot-backend/api/chat.py
@@ -22,14 +22,19 @@ from uuid import uuid4
 
 from fastapi import APIRouter, Body, Depends, HTTPException, Request
 from fastapi.responses import JSONResponse, StreamingResponse
-from pydantic import BaseModel, Field
-
 from auth_middleware import get_current_user
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from autobot_shared.error_utils import safe_http_detail
 from autobot_shared.time_utils import parse_utc_iso, utc_timestamp
-from constants.threshold_constants import CategoryDefaults, TimingConstants
+from constants.threshold_constants import TimingConstants
 
+from api.schemas_chat import (
+    ChatMessage,
+    ChatPreferences,
+    DetectLanguageRequest,
+    EnhancedChatMessage,
+    TranslateRequest,
+)
 from api.schemas_common import DataResponse
 
 # Import dependencies and utilities - Using available dependencies
@@ -313,108 +318,6 @@ logger = logging.getLogger(__name__)
 from api.chat_sessions import router as sessions_router
 
 router.include_router(sessions_router)
-
-# ====================================================================
-# Request/Response Models
-# ====================================================================
-
-
-class ChatMessage(BaseModel):
-    """Chat message model for requests"""
-
-    content: str = Field(
-        ..., min_length=1, max_length=50000, description="Message content"
-    )
-    role: str = Field(
-        default=CategoryDefaults.ROLE_USER,
-        pattern="^(user|assistant|system)$",
-        description="Message role",
-    )
-    session_id: Optional[str] = Field(None, description="Chat session ID")
-    message_type: Optional[str] = Field("text", description="Message type")
-    metadata: Optional[Metadata] = Field(
-        default_factory=dict, description="Additional metadata"
-    )
-    language: Optional[str] = Field(
-        None,
-        description="Preferred response language code (e.g. 'en', 'es', 'de'). "
-        "Overrides personality language when set.",
-    )
-
-
-class ChatResponse(BaseModel):
-    """Chat response model"""
-
-    content: str
-    role: str = CategoryDefaults.ROLE_ASSISTANT
-    session_id: str
-    message_id: str
-    timestamp: datetime
-    metadata: Metadata = Field(default_factory=dict)
-
-
-class MessageHistory(BaseModel):
-    """Message history response model"""
-
-    messages: List[Metadata]
-    session_id: str
-    total_count: int
-    page: int = 1
-    per_page: int = 50
-
-
-class EnhancedChatMessage(BaseModel):
-    """Enhanced chat message model with AI Stack integration (Issue #708 consolidation)."""
-
-    content: str = Field(
-        ..., min_length=1, max_length=50000, description="Message content"
-    )
-    role: str = Field(
-        default=CategoryDefaults.ROLE_USER,
-        pattern="^(user|assistant|system)$",
-        description="Message role",
-    )
-    session_id: Optional[str] = Field(None, description="Chat session ID")
-    message_type: Optional[str] = Field("text", description="Message type")
-    metadata: Optional[Metadata] = Field(
-        default_factory=dict, description="Additional metadata"
-    )
-
-    language: Optional[str] = Field(
-        None,
-        description="Preferred response language code (e.g. 'en', 'es', 'de'). "
-        "Overrides personality language when set.",
-    )
-
-    # AI Stack specific fields
-    use_ai_stack: bool = Field(
-        True, description="Whether to use AI Stack for enhanced responses"
-    )
-    use_knowledge_base: bool = Field(
-        True, description="Whether to include knowledge base context"
-    )
-    response_style: str = Field(
-        "conversational", description="Response style preference"
-    )
-    include_sources: bool = Field(
-        True, description="Whether to include source citations"
-    )
-
-
-class ChatPreferences(BaseModel):
-    """Chat preferences for customizing AI behavior (Issue #708 consolidation)."""
-
-    response_length: str = Field(
-        "medium", description="Preferred response length (short, medium, long)"
-    )
-    technical_level: str = Field("adaptive", description="Technical complexity level")
-    include_reasoning: bool = Field(
-        False, description="Include reasoning steps in responses"
-    )
-    fact_checking: bool = Field(
-        True, description="Enable fact checking against knowledge base"
-    )
-
 
 # ====================================================================
 # Configuration and State Management
@@ -2240,38 +2143,6 @@ async def get_enhanced_chat_capabilities(
 # ====================================================================
 # Issue #1328: Translation Shortcut Endpoint
 # ====================================================================
-
-
-class TranslateRequest(BaseModel):
-    """Request model for direct translation."""
-
-    text: str = Field(
-        ...,
-        min_length=1,
-        max_length=50000,
-        description="Text to translate",
-    )
-    target_language: str = Field(
-        ...,
-        min_length=1,
-        max_length=50,
-        description="Target language name",
-    )
-    source_language: Optional[str] = Field(
-        None,
-        description="Source language (auto-detect if omitted)",
-    )
-
-
-class DetectLanguageRequest(BaseModel):
-    """Request model for language detection."""
-
-    text: str = Field(
-        ...,
-        min_length=1,
-        max_length=50000,
-        description="Text to detect language of",
-    )
 
 
 @with_error_handling(
